Insider_threat_detection/requirements_new.txt
Flask>=2.0.0
mysql-connector-python>=8.0.0
scikit-learn>=1.0.0
numpy>=1.20.0
pandas>=1.3.0
joblib>=1.0.0
bcrypt>=3.2.0
Werkzeug>=2.0.0
watchdog>=2.0.0
psutil>=5.8.0
pywin32>=300